refactor: useSearchQuery 훅 생성
