+++ src/index.tsx
@@ -63,9 +63,15 @@ body {
 `;
 
 const root = ReactDOM.createRoot(document.getElementById('root') as HTMLElement);
+// root.render(
+//   <React.StrictMode>
+//     <GlobalStyle />
+//     <App />
+//   </React.StrictMode>
+// );
 root.render(
-  <React.StrictMode>
+  <>
     <GlobalStyle />
     <App />
-  </React.StrictMode>
+  </>
 );
